+++ banyand/tsdb/buffer.go
@@ -25,6 +25,7 @@ import (
 
 	"github.com/dgraph-io/badger/v3/skl"
 	"github.com/dgraph-io/badger/v3/y"
+	"github.com/pkg/errors"
 
 	"github.com/apache/skywalking-banyandb/api/common"
 	"github.com/apache/skywalking-banyandb/banyand/observability"
@@ -32,11 +33,13 @@ import (
 	"github.com/apache/skywalking-banyandb/pkg/logger"
 	"github.com/apache/skywalking-banyandb/pkg/meter"
 	"github.com/apache/skywalking-banyandb/pkg/run"
+	"github.com/apache/skywalking-banyandb/pkg/wal"
 )
 
 const (
-	defaultSize = 1 << 20 // 1MB
-	nodeAlign   = int(unsafe.Sizeof(uint64(0))) - 1
+	defaultSize        = 1 << 20 // 1MB
+	nodeAlign          = int(unsafe.Sizeof(uint64(0))) - 1
+	defaultWalSyncMode = false
 )
 
 var (
@@ -53,13 +56,15 @@ func init() {
 }
 
 type operation struct {
-	key   []byte
-	value []byte
-	epoch uint64
+	recoveryDoneFn func()
+	key            []byte
+	value          []byte
+	epoch          uint64
 }
 
 type flushEvent struct {
-	data *skl.Skiplist
+	data         *skl.Skiplist
+	walSegmentID wal.SegmentID
 }
 
 type onFlush func(shardIndex int, skl *skl.Skiplist) error
@@ -71,12 +76,15 @@ type bufferShardBucket struct {
 	writeWaitGroup   *sync.WaitGroup
 	flushWaitGroup   *sync.WaitGroup
 	log              *logger.Logger
+	wal              wal.WAL
 	immutables       []*skl.Skiplist
 	labelValues      []string
 	shardLabelValues []string
 	index            int
 	capacity         int
 	mutex            sync.RWMutex
+	walSyncMode      bool
+	enableWal        bool
 }
 
 // Buffer is an exported struct that represents a buffer composed of multiple shard buckets.
@@ -88,11 +96,17 @@ type Buffer struct {
 	writeWaitGroup sync.WaitGroup
 	flushWaitGroup sync.WaitGroup
 	numShards      int
+	enableWal      bool
 	closerOnce     sync.Once
 }
 
 // NewBuffer creates a new Buffer instance with the given parameters.
 func NewBuffer(log *logger.Logger, position common.Position, flushSize, writeConcurrency, numShards int, onFlushFn onFlush) (*Buffer, error) {
+	return NewBufferWithWal(log, position, flushSize, writeConcurrency, numShards, onFlushFn, false, nil)
+}
+
+// NewBufferWithWal creates a new Buffer instance with the given parameters.
+func NewBufferWithWal(log *logger.Logger, position common.Position, flushSize, writeConcurrency, numShards int, onFlushFn onFlush, enableWal bool, walPath *string) (*Buffer, error) {
 	buckets := make([]bufferShardBucket, numShards)
 	buffer := &Buffer{
 		buckets:     buckets,
@@ -100,6 +114,7 @@ func NewBuffer(log *logger.Logger, position common.Position, flushSize, writeCon
 		onFlushFn:   onFlushFn,
 		entryCloser: run.NewCloser(1),
 		log:         log.Named("buffer"),
+		enableWal:   enableWal,
 	}
 	buffer.writeWaitGroup.Add(numShards)
 	buffer.flushWaitGroup.Add(numShards)
@@ -115,17 +130,27 @@ func NewBuffer(log *logger.Logger, position common.Position, flushSize, writeCon
 			log:              buffer.log.Named(fmt.Sprintf("shard-%d", i)),
 			labelValues:      append(position.LabelValues(), fmt.Sprintf("%d", i)),
 			shardLabelValues: position.ShardLabelValues(),
+			enableWal:        enableWal,
 		}
 		buckets[i].start(onFlushFn)
+		if enableWal {
+			if walPath == nil {
+				return nil, errors.New("wal path is required")
+			}
+			shardWalPath := fmt.Sprintf("%s/shard-%d", *walPath, i)
+			if err := buckets[i].startWal(shardWalPath, defaultWalSyncMode); err != nil {
+				return nil, errors.Wrap(err, "failed to start wal")
+			}
+		}
 		maxBytes.Set(float64(flushSize), buckets[i].labelValues...)
 	}
 	return buffer, nil
 }
 
 // Write adds a key-value pair with a timestamp to the appropriate shard bucket in the buffer.
-func (b *Buffer) Write(key, value []byte, timestamp time.Time) {
+func (b *Buffer) Write(key, value []byte, timestamp time.Time) error {
 	if b == nil || !b.entryCloser.AddRunning() {
-		return
+		return errors.New("buffer is invalid")
 	}
 	defer b.entryCloser.Done()
 	index := b.getShardIndex(key)
@@ -133,7 +158,15 @@ func (b *Buffer) Write(key, value []byte, timestamp time.Time) {
 		b.log.Debug().Uint64("shard", index).Bytes("key", key).
 			Time("ts", timestamp).Msg("route a shard")
 	}
+
+	if b.enableWal {
+		if err := b.buckets[index].writeWal(key, value, timestamp); err != nil {
+			return errors.Wrap(err, "failed to write wal")
+		}
+	}
+
 	b.buckets[index].writeCh <- operation{key: key, value: value, epoch: uint64(timestamp.UnixNano())}
+	return nil
 }
 
 // Read retrieves the value associated with the given key and timestamp from the appropriate shard bucket in the buffer.
@@ -179,6 +212,11 @@ func (b *Buffer) Close() error {
 		}
 		for i := 0; i < b.numShards; i++ {
 			close(b.buckets[i].flushCh)
+			if b.enableWal {
+				if err := b.buckets[i].wal.Close(); err != nil {
+					b.buckets[i].log.Err(err).Msg("closing buffer shard wal failed")
+				}
+			}
 		}
 		b.flushWaitGroup.Wait()
 	})
@@ -218,13 +256,24 @@ func (bsb *bufferShardBucket) start(onFlushFn onFlush) {
 		for event := range bsb.flushCh {
 			oldSkipList := event.data
 			memSize := oldSkipList.MemSize()
+			onFlushFnDone := false
 			t1 := time.Now()
 			for {
-				if err := onFlushFn(bsb.index, oldSkipList); err != nil {
-					bsb.log.Err(err).Msg("flushing immutable buffer failed. Retrying...")
-					flushNum.Inc(1, append(bsb.labelValues[:2], "true")...)
-					time.Sleep(time.Second)
-					continue
+				if !onFlushFnDone {
+					if err := onFlushFn(bsb.index, oldSkipList); err != nil {
+						bsb.log.Err(err).Msg("flushing immutable buffer failed. Retrying...")
+						flushNum.Inc(1, append(bsb.labelValues[:2], "true")...)
+						time.Sleep(time.Second)
+						continue
+					}
+					onFlushFnDone = true
+				}
+				if bsb.enableWal {
+					if err := bsb.wal.Delete(event.walSegmentID); err != nil {
+						bsb.log.Err(err).Msg("delete wal segment file failed. Retrying...")
+						time.Sleep(time.Second)
+						continue
+					}
 				}
 				break
 			}
@@ -249,23 +298,40 @@ func (bsb *bufferShardBucket) start(onFlushFn onFlush) {
 			volume += len(k) + int(v.EncodedSize()) + skl.MaxNodeSize + nodeAlign
 			memSize := bsb.mutable.MemSize()
 			mutableBytes.Set(float64(memSize), bsb.labelValues...)
-			if volume >= bsb.capacity || memSize >= int64(bsb.capacity) {
-				bsb.triggerFlushing()
-				volume = 0
+			if bsb.enableWal && op.recoveryDoneFn != nil {
+				// If it's a recovery write, we don't need to trigger the flushing.
+			} else if volume >= bsb.capacity || memSize >= int64(bsb.capacity) {
+				if err := bsb.triggerFlushing(); err != nil {
+					bsb.log.Err(err).Msg("triggering flushing failed")
+				} else {
+					volume = 0
+				}
 			}
 			bsb.mutable.Put(k, v)
+			if bsb.enableWal && op.recoveryDoneFn != nil {
+				op.recoveryDoneFn()
+			}
 		}
 	}()
 }
 
-func (bsb *bufferShardBucket) triggerFlushing() {
+func (bsb *bufferShardBucket) triggerFlushing() error {
+	var walSegmentID wal.SegmentID
+	if bsb.enableWal {
+		segment, err := bsb.wal.Rotate()
+		if err != nil {
+			return errors.Wrap(err, "rotating wal failed")
+		}
+		walSegmentID = segment.GetSegmentID()
+	}
+
 	for {
 		select {
-		case bsb.flushCh <- flushEvent{data: bsb.mutable}:
+		case bsb.flushCh <- flushEvent{data: bsb.mutable, walSegmentID: walSegmentID}:
 			bsb.mutex.Lock()
 			defer bsb.mutex.Unlock()
 			bsb.swap()
-			return
+			return nil
 		default:
 		}
 		time.Sleep(10 * time.Second)
@@ -275,4 +341,108 @@ func (bsb *bufferShardBucket) triggerFlushing() {
 func (bsb *bufferShardBucket) swap() {
 	bsb.immutables = append(bsb.immutables, bsb.mutable)
 	bsb.mutable = skl.NewSkiplist(int64(bsb.capacity))
+}
+
+func (bsb *bufferShardBucket) startWal(path string, syncMode bool) error {
+	wal, err := wal.New(path, wal.DefaultOptions)
+	if err != nil {
+		return errors.Wrap(err, fmt.Sprintf("failed to create wal: %s", path))
+	}
+	bsb.wal = wal
+	bsb.walSyncMode = syncMode
+	bsb.log.Info().Msg(fmt.Sprintf(
+		"wal started with path: %s, sync mode: %v", path, syncMode))
+
+	if err = bsb.recoveryWal(); err != nil {
+		return errors.Wrap(err, "failed to recovery wal")
+	}
+	return nil
+}
+
+func (bsb *bufferShardBucket) recoveryWal() error {
+	segments, err := bsb.wal.ReadAllSegments()
+	if err != nil {
+		return errors.Wrap(err, "failed to load wal segments")
+	}
+
+	recoveredRecords := 0
+	for index, segment := range segments {
+		recoveredRecords += len(segment.GetLogEntries())
+		isWorkSegment := index == len(segments)-1
+		if isWorkSegment {
+			bsb.recoveryWorkSegment(segment)
+		} else {
+			bsb.recoveryStableSegment(segment)
+		}
+		bsb.log.Info().Msg(fmt.Sprintf(
+			"recovered %d log records from wal segment %d",
+			len(segment.GetLogEntries()),
+			segment.GetSegmentID()))
+	}
+	bsb.log.Info().Msg(fmt.Sprintf(
+		"recovered %d log records from wal", recoveredRecords))
+	return nil
+}
+
+func (bsb *bufferShardBucket) recoveryWorkSegment(segment wal.Segment) {
+	var wg sync.WaitGroup
+	wg.Add(len(segment.GetLogEntries()))
+	for _, logEntry := range segment.GetLogEntries() {
+		timestamps := logEntry.GetTimestamps()
+		values := logEntry.GetValues()
+		elementIndex := 0
+		for element := values.Front(); element != nil; element = element.Next() {
+			timestamp := timestamps[elementIndex]
+			bsb.writeCh <- operation{
+				key:   logEntry.GetSeriesID(),
+				value: element.Value.([]byte),
+				epoch: uint64(timestamp.UnixNano()),
+				recoveryDoneFn: func() {
+					wg.Done()
+					if bsb.log.Trace().Enabled() {
+						bsb.log.Trace().Msg(fmt.Sprintf("recovered key: %v, ts: %v",
+							logEntry.GetSeriesID(), timestamp.UnixNano()))
+					}
+				},
+			}
+			elementIndex++
+		}
+	}
+	wg.Wait()
+}
+
+func (bsb *bufferShardBucket) recoveryStableSegment(segment wal.Segment) {
+	for _, logEntries := range segment.GetLogEntries() {
+		timestamps := logEntries.GetTimestamps()
+		values := logEntries.GetValues()
+		elementIndex := 0
+		for element := values.Front(); element != nil; element = element.Next() {
+			timestamp := timestamps[elementIndex]
+			k := y.KeyWithTs(logEntries.GetSeriesID(), uint64(timestamp.UnixNano()))
+			v := y.ValueStruct{Value: element.Value.([]byte)}
+			bsb.mutable.Put(k, v)
+			elementIndex++
+		}
+	}
+	bsb.flushCh <- flushEvent{data: bsb.mutable, walSegmentID: segment.GetSegmentID()}
+	// Sync recover data to immutables
+	bsb.swap()
+}
+
+func (bsb *bufferShardBucket) writeWal(key, value []byte, timestamp time.Time) error {
+	if !bsb.walSyncMode {
+		bsb.wal.Write(key, timestamp, value, nil)
+		return nil
+	}
+
+	var walErr error
+	var wg sync.WaitGroup
+	wg.Add(1)
+	walCallback := func(key []byte, t time.Time, value []byte, err error) {
+		walErr = err
+		wg.Done()
+	}
+	bsb.wal.Write(key, timestamp, value, walCallback)
+	wg.Wait()
+	return walErr
 }